requirements.txt
bottle
python-dotenv
stripe
pillow